Adding product routes, registration and login with jwt authentication

+++ src/main/java/com/bdweb/frumeterapi/controller/ProductController.java
@@ -5,6 +5,7 @@ import com.bdweb.frumeterapi.model.dto.ProductDTO;
 import com.bdweb.frumeterapi.repository.ProductRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
@@ -38,11 +39,11 @@ public class ProductController {
 
     @Transactional
     @PostMapping("/save")
-    public String saveProduct(@RequestBody ProductDTO productDTO) {
+    public ResponseEntity saveProduct(@RequestBody ProductDTO productDTO) {
         Product product = new Product(productDTO);
         try {
             productRepository.save(product);
-            return "Produto salvo com sucesso!";
+            return new ResponseEntity<>("Produto salvo com sucesso!", HttpStatus.OK);
         } catch (Exception exception) {
             log.error("Erro ao salvar produto", exception);
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Erro ao cadastrar produto, por favor tente novamente.");
@@ -50,10 +51,14 @@ public class ProductController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteProduct(@PathVariable("id") Long id) {
+    public ResponseEntity deleteProduct(@PathVariable("id") Long id) {
         try {
-            productRepository.deleteById(id);
-            return "Produto excluído com sucesso!";
+            if (productRepository.findById(id).isPresent()) {
+                productRepository.deleteById(id);
+                return new ResponseEntity<>("Produto excluído com sucesso!", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Nenhum produto encontrado com esse id", HttpStatus.NOT_FOUND);
+            }
         } catch (Exception exception) {
             log.error("Erro ao excluir produto", exception);
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Erro ao excluir produto, por favor tente novamente.");
@@ -61,13 +66,13 @@ public class ProductController {
     }
 
     @PutMapping("/update")
-    public String putProduct(@RequestBody Product product) {
+    public ResponseEntity putProduct(@RequestBody Product product) {
         try {
             if (productRepository.findById(product.getId()).isPresent()) {
                 productRepository.save(product);
-                return "Produto atualizado com sucesso!";
+                return new ResponseEntity<>("Produto atualizado com sucesso!", HttpStatus.OK);
             } else {
-                return "Não foi encontrado nenhum produto com o id informado.";
+                return new ResponseEntity<>("Não foi encontrado nenhum produto com o id informado.", HttpStatus.OK);
             }
         } catch (Exception exception) {
             log.error("Erro ao atualizar produto", exception);
